code/requirements.txt
# requirements.txt
langchain==0.2.0
langchain-openai==0.1.0
langchain-google-vertexai==1.0.0
amplpy==0.4.9
gurobipy==11.0.0
openai==1.0.0
google-auth==2.0.0
google-auth-transport-requests==0.1.0
requests==2.31.0
argparse==1.4.0
json==2.0.9
subprocess==0.1.0
shutil==1.7.0
re==2.2.1
time==3.9.0
csv==1.0
textwrap==1.0
sys==3.9.0